django/requirements.txt
Django==4.2.0
django-cors-headers>=4.2.0
djangorestframework>=3.11,<4.0
psycopg2-binary>=2.8
redis>=4.0
requests>=2.31.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0